Refactor model deployment listing to FoundryService
Moved the Azure model deployment listing logic from TeamService to FoundryService for better separation of concerns and code reuse. Updated TeamService to use FoundryService for model deployment validation and status summary. Improved logging and configuration handling in FoundryService.
