+++ src/backend/v3/common/services/foundry_service.py
@@ -1,7 +1,8 @@
 from typing import Any, Dict
-
+import logging
 from azure.ai.projects.aio import AIProjectClient
-
+from git import List
+import aiohttp
 from common.config.app_config import config
 
 
@@ -15,6 +16,12 @@ class FoundryService:
 
     def __init__(self, client: AIProjectClient | None = None) -> None:
         self._client = client
+        self.logger = logging.getLogger(__name__)
+        # Model validation configuration
+        self.subscription_id = config.AZURE_AI_SUBSCRIPTION_ID
+        self.resource_group = config.AZURE_AI_RESOURCE_GROUP
+        self.project_name = config.AZURE_AI_PROJECT_NAME
+        self.project_endpoint = config.AZURE_AI_PROJECT_ENDPOINT
 
     async def get_client(self) -> AIProjectClient:
         if self._client is None:
@@ -31,3 +38,62 @@ class FoundryService:
         client = await self.get_client()
         conn = await client.connections.get(name=name)
         return conn.as_dict() if hasattr(conn, "as_dict") else dict(conn)
+
+    # -----------------------
+    # Model validation methods
+    # -----------------------
+
+    async def list_model_deployments(self) -> List[Dict[str, Any]]:
+        """
+        List all model deployments in the Azure AI project using the REST API.
+        """
+        if not all([self.subscription_id, self.resource_group, self.project_name]):
+            self.logger.error("Azure AI project configuration is incomplete")
+            return []
+
+        try:
+            token = await config.get_access_token()
+
+            url = (
+                f"https://management.azure.com/subscriptions/{self.subscription_id}/"
+                f"resourceGroups/{self.resource_group}/providers/Microsoft.MachineLearningServices/"
+                f"workspaces/{self.project_name}/onlineEndpoints"
+            )
+
+            headers = {
+                "Authorization": f"Bearer {token}",
+                "Content-Type": "application/json",
+            }
+            params = {"api-version": "2024-10-01"}
+
+            async with aiohttp.ClientSession() as session:
+                async with session.get(url, headers=headers, params=params) as response:
+                    if response.status == 200:
+                        data = await response.json()
+                        deployments = data.get("value", [])
+                        deployment_info: List[Dict[str, Any]] = []
+                        for deployment in deployments:
+                            deployment_info.append(
+                                {
+                                    "name": deployment.get("name"),
+                                    "model": deployment.get("properties", {}).get(
+                                        "model", {}
+                                    ),
+                                    "status": deployment.get("properties", {}).get(
+                                        "provisioningState"
+                                    ),
+                                    "endpoint_uri": deployment.get(
+                                        "properties", {}
+                                    ).get("scoringUri"),
+                                }
+                            )
+                        return deployment_info
+                    else:
+                        error_text = await response.text()
+                        self.logger.error(
+                            f"Failed to list deployments. Status: {response.status}, Error: {error_text}"
+                        )
+                        return []
+        except Exception as e:
+            self.logger.error(f"Error listing model deployments: {e}")
+            return []
